requirements_local.txt
# =============================================================================
# OpenEnv Education Track — Local CPU Requirements (Demo)
# =============================================================================
# Install:  pip install -r requirements_local.txt
# CPU-only PyTorch (~200MB instead of ~2.5GB with CUDA)
# =============================================================================

--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.3.0
torchvision>=0.18.0

# HuggingFace
transformers>=4.43.0
trl>=0.9.0
accelerate>=0.32.0
peft>=0.12.0
datasets>=2.20.0
huggingface_hub>=0.24.0
tokenizers>=0.19.0
sentencepiece>=0.2.0

# Environment servers
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0

# Utilities
scipy>=1.13.0
numpy>=1.26.0
requests>=2.31.0
openai>=1.0.0
python-dotenv>=1.0.0
